fix: 修复路网线宽类型转换错误

+++ app/src/main/java/com/chasegame/lastfreedom/RoadOverlayManager.kt
@@ -53,7 +53,7 @@ class RoadOverlayManager(
         }
         // 动态调整路网线宽：zoom 18 → 15f，zoom 13 → 0f（线性插值）
         if (isVisible) {
-            val strokeW = ((zoom - MIN_ZOOM_FOR_ROADS) / (18.0 - MIN_ZOOM_FOR_ROADS) * 15.0f).coerceIn(1f, 15f)
+            val strokeW = ((zoom - MIN_ZOOM_FOR_ROADS) / (18.0 - MIN_ZOOM_FOR_ROADS) * 15.0).toFloat().coerceIn(1f, 15f)
             roadPolylines.forEach { it.outlinePaint.strokeWidth = strokeW }
             map.invalidate()
         }
